FIN-28 Receivables refactoring - making StickyDrawer component. using it inside Payables, Receivables
