+++ src/src_V2/modules/manager/financial/components/index.js
@@ -5,3 +5,4 @@ export {
   TooltipTableCell,
   TableCellWrapper,
 } from './sticky-table/StickyTable';
+export { default as StickyDrawer } from './sticky-drawer/StickyDrawer';
