+++ src/src_V2/modules/manager/financial/components/sticky-drawer/StickyDrawer.module.css
@@ -1,5 +1,5 @@
-.mainText {
-  font-size: 17px;
+.contentDiv {
+  width: 55vw;
 }
 
 .mainBox {
@@ -13,27 +13,13 @@
 
 .layout {
   min-height: 82vh;
-  margin-right: 25px;
-  margin-left: 25px;
-}
-
-.layoutDistance {
-  margin-top: 30px;
-  margin-bottom: 30px;
-}
-
-.textDiv {
-  margin-left: 25px;
-}
-
-.receivableId {
-  color: var(--darkGrey);
-  font-weight: 600;
-}
-
-.supplierName {
-  color: var(--secondary);
-  font-size: 12px;
+  padding-right: 25px;
+  padding-left: 25px;
+  max-height: calc(99.8vh - 156px);
+  overflow-y: auto;
+  text-align: justify;
+  overflow-x: hidden;
+  box-sizing: border-box;
 }
 
 .closeBtnDiv {
@@ -49,8 +35,11 @@
 }
 
 .footerBox {
+  position: absolute;
+  bottom: 0;
   width: 100%;
   justify-content: end;
+  background-color: var(--white);
 }
 
 .btnDiv {
