+++ src/src_V2/modules/manager/financial/mock/receivable.js
@@ -97,7 +97,7 @@ const receivablesList = () => [
     440,
     true,
     23000,
-    1,
+    2,
     '2021-12-26T09:36:19Z',
     'Mon Jan 03 2022 16:31:13 GMT+0400 (Armenia Standard Time)',
     true,
@@ -131,7 +131,7 @@ const receivablesList = () => [
     440,
     true,
     23000,
-    1,
+    4,
     '2021-12-26T09:36:19Z',
     'Mon Jan 03 2022 16:31:13 GMT+0400 (Armenia Standard Time)',
     true,
@@ -217,7 +217,7 @@ const receivablesList = () => [
     800,
     false,
     2500,
-    1,
+    3,
     '2021-12-26T09:36:19Z',
     'Mon Jan 03 2022 16:31:13 GMT+0400 (Armenia Standard Time)',
     true,
@@ -435,7 +435,7 @@ const receivablesList = () => [
     440,
     false,
     23000,
-    1,
+    4,
     '2021-12-26T09:36:19Z',
     'Mon Jan 03 2022 16:31:13 GMT+0400 (Armenia Standard Time)',
     false,
