+++ src/src_V2/modules/manager/financial/payables/edit-drawer/notes/Notes.jsx
