+++ src/src_V2/modules/manager/financial/payables/edit-drawer/tour-details/TourDetails.jsx
@@ -8,8 +8,8 @@ import {
   Typography,
 } from '@mui/material';
 
-import { generateDate } from '../../../../../../../utils';
-import { PaymentStatus } from '../../../../constants';
+import { generateDate } from '../../../../../../utils';
+import { PaymentStatus } from '../../../constants';
 import commonStyles from '../style.module.css';
 import styles from './TourDetails.module.css';
