+++ src/src_V2/modules/manager/financial/payables/edit-drawer/tour-details/TourDetails.module.css
