--- src/src_V2/modules/manager/financial/payables/edit/drawer/EditDrawer.module.css
@@ -1,3 +1,0 @@
-.contentDiv {
-  width: 55vw;
-}
