+++ src/src_V2/modules/manager/financial/receivables/edit-drawer/editable-info/EditableInfo.jsx
@@ -1,4 +1,4 @@
-import { DateInput } from '../../../../components';
+import { DateInput } from '../../../components';
 import commonStyles from '../style.module.css';
 
 export default function EditableInfo({ values, setFieldValue }) {
