+++ src/src_V2/modules/manager/financial/receivables/edit-drawer/general-info/GeneralInfo.module.css
