+++ src/src_V2/modules/manager/financial/receivables/edit-drawer/notes/Notes.module.css
