+++ src/src_V2/modules/manager/financial/receivables/edit-drawer/payment-history/PaymentHistory.module.css
@@ -39,3 +39,7 @@
   display: flex;
   align-items: flex-start;
 }
+
+.uploadInput {
+  display: none;
+}
