+++ src/src_V2/modules/manager/financial/receivables/edit-drawer/style.module.css
@@ -39,5 +39,5 @@
 /* DatePicker */
 
 .dueDatePickerInput > div > div {
-  width: 52.5vw;
+  width: 51.5vw;
 }
