+++ src/src_V2/modules/manager/financial/receivables/edit-drawer/tour-details/TourDetails.module.css
