--- src/src_V2/modules/manager/financial/receivables/edit/drawer/EditDrawer.module.css
@@ -1,3 +1,0 @@
-.contentDiv {
-  width: 55vw;
-}
