+++ src/src_V2/modules/manager/financial/receivables/table/ReceivableTable.jsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { StickyTable } from '../../components';
 import { receivableColumns } from '../../constants';
 import { getComparator } from '../../utils';
-import EditDrawer from '../edit/drawer/EditDrawer';
+import EditDrawer from '../edit-drawer/EditDrawer';
 
 export default function ReceivableTable({ receivables, receivablesLoading }) {
   const [drawerState, setDrawerState] = useState({
